User Space Stack Pointer Initialization edited and compiles without warning

+++ Core/Src/main.c
@@ -73,7 +73,7 @@ int main(void)
 	
 	/*User Space Stack Pointer Initialization*/
 	extern const uint32_t _eustack[];
-	__set_PSP(_eustack);
+	__set_PSP((uint32_t)_eustack);
 
 	/* USER CODE BEGIN WHILE */
 	kready = 1;
